initial commit. fixes test data path. added pre-commit and build workflow. incomplete work skip ci.
