+++ .pre-commit-config.yaml
@@ -1,0 +1,10 @@
+repos:
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.15.0
+    hooks:
+      # Run the linter.
+      - id: ruff-check
+        args: [--verbose, --fix]
+      # Run the formatter.
+      - id: ruff-format
+        args: [--verbose]
