+++ tests/test_files/6eey/6eey_final.pdb
